requirements.txt
django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-environ==0.12.0
pytz==2025.1
sqlparse==0.5.3
jsonfield==3.1.0